removes HybridCodecPipeline
