+++ src/zarr/codecs/__init__.py
@@ -7,4 +7,4 @@ from zarr.codecs.gzip import GzipCodec  # noqa: F401
 from zarr.codecs.sharding import ShardingCodec, ShardingCodecIndexLocation  # noqa: F401
 from zarr.codecs.transpose import TransposeCodec  # noqa: F401
 from zarr.codecs.zstd import ZstdCodec  # noqa: F401
-from zarr.codecs.pipeline import CodecPipeline, BatchedCodecPipeline, HybridCodecPipeline  # noqa: F401
+from zarr.codecs.batched_codec_pipeline import BatchedCodecPipeline  # noqa: F401
