+++ src/zarr/codecs/batched_codec_pipeline.py
@@ -1,0 +1,446 @@
+from __future__ import annotations
+
+from itertools import islice
+from typing import TYPE_CHECKING, Iterator, TypeVar, Iterable
+from warnings import warn
+import numpy as np
+from dataclasses import dataclass
+
+from zarr.config import config
+from zarr.abc.codec import (
+    ByteGetter,
+    ByteSetter,
+    Codec,
+    CodecPipeline,
+    ArrayArrayCodec,
+    ArrayBytesCodec,
+    ArrayBytesCodecPartialDecodeMixin,
+    ArrayBytesCodecPartialEncodeMixin,
+    BytesBytesCodec,
+)
+from zarr.codecs.registry import get_codec_class
+from zarr.common import JSON, concurrent_map, parse_named_configuration
+from zarr.indexing import is_total_slice
+from zarr.metadata import ArrayMetadata
+
+if TYPE_CHECKING:
+    from typing_extensions import Self
+    from zarr.common import ArraySpec, BytesLike, SliceSelection
+
+T = TypeVar("T")
+U = TypeVar("U")
+
+
+def _unzip2(iterable: Iterable[tuple[T, U]]) -> tuple[list[T], list[U]]:
+    out0: list[T] = []
+    out1: list[U] = []
+    for item0, item1 in iterable:
+        out0.append(item0)
+        out1.append(item1)
+    return (out0, out1)
+
+
+def batched(iterable: Iterable[T], n: int) -> Iterable[tuple[T, ...]]:
+    if n < 1:
+        raise ValueError("n must be at least one")
+    it = iter(iterable)
+    while batch := tuple(islice(it, n)):
+        yield batch
+
+
+def resolve_batched(codec: Codec, chunk_specs: Iterable[ArraySpec]) -> Iterable[ArraySpec]:
+    return [codec.resolve_metadata(chunk_spec) for chunk_spec in chunk_specs]
+
+
+@dataclass(frozen=True)
+class BatchedCodecPipeline(CodecPipeline):
+    array_array_codecs: tuple[ArrayArrayCodec, ...]
+    array_bytes_codec: ArrayBytesCodec
+    bytes_bytes_codecs: tuple[BytesBytesCodec, ...]
+    batch_size: int
+
+    @classmethod
+    def from_dict(cls, data: Iterable[JSON | Codec], *, batch_size: int | None = None) -> Self:
+        out: list[Codec] = []
+        if not isinstance(data, Iterable):
+            raise TypeError(f"Expected iterable, got {type(data)}")
+
+        for c in data:
+            if isinstance(c, Codec):
+                out.append(c)
+            else:
+                name_parsed, _ = parse_named_configuration(c, require_configuration=False)
+                out.append(get_codec_class(name_parsed).from_dict(c))  # type: ignore[arg-type]
+        return cls.from_list(out, batch_size=batch_size)
+
+    def to_dict(self) -> JSON:
+        return [c.to_dict() for c in self]
+
+    def evolve(self, array_spec: ArraySpec) -> Self:
+        return type(self).from_list([c.evolve(array_spec) for c in self])
+
+    @staticmethod
+    def codecs_from_list(
+        codecs: list[Codec],
+    ) -> tuple[tuple[ArrayArrayCodec, ...], ArrayBytesCodec, tuple[BytesBytesCodec, ...]]:
+        from zarr.codecs.sharding import ShardingCodec
+
+        if not any(isinstance(codec, ArrayBytesCodec) for codec in codecs):
+            raise ValueError("Exactly one array-to-bytes codec is required.")
+
+        prev_codec: Codec | None = None
+        for codec in codecs:
+            if prev_codec is not None:
+                if isinstance(codec, ArrayBytesCodec) and isinstance(prev_codec, ArrayBytesCodec):
+                    raise ValueError(
+                        f"ArrayBytesCodec '{type(codec)}' cannot follow after "
+                        + f"ArrayBytesCodec '{type(prev_codec)}' because exactly "
+                        + "1 ArrayBytesCodec is allowed."
+                    )
+                if isinstance(codec, ArrayBytesCodec) and isinstance(prev_codec, BytesBytesCodec):
+                    raise ValueError(
+                        f"ArrayBytesCodec '{type(codec)}' cannot follow after "
+                        + f"BytesBytesCodec '{type(prev_codec)}'."
+                    )
+                if isinstance(codec, ArrayArrayCodec) and isinstance(prev_codec, ArrayBytesCodec):
+                    raise ValueError(
+                        f"ArrayArrayCodec '{type(codec)}' cannot follow after "
+                        + f"ArrayBytesCodec '{type(prev_codec)}'."
+                    )
+                if isinstance(codec, ArrayArrayCodec) and isinstance(prev_codec, BytesBytesCodec):
+                    raise ValueError(
+                        f"ArrayArrayCodec '{type(codec)}' cannot follow after "
+                        + f"BytesBytesCodec '{type(prev_codec)}'."
+                    )
+            prev_codec = codec
+
+        if any(isinstance(codec, ShardingCodec) for codec in codecs) and len(codecs) > 1:
+            warn(
+                "Combining a `sharding_indexed` codec disables partial reads and "
+                + "writes, which may lead to inefficient performance."
+            )
+
+        return (
+            tuple(codec for codec in codecs if isinstance(codec, ArrayArrayCodec)),
+            next(codec for codec in codecs if isinstance(codec, ArrayBytesCodec)),
+            tuple(codec for codec in codecs if isinstance(codec, BytesBytesCodec)),
+        )
+
+    @classmethod
+    def from_list(cls, codecs: list[Codec], *, batch_size: int | None = None) -> Self:
+        array_array_codecs, array_bytes_codec, bytes_bytes_codecs = cls.codecs_from_list(codecs)
+
+        return cls(
+            array_array_codecs=array_array_codecs,
+            array_bytes_codec=array_bytes_codec,
+            bytes_bytes_codecs=bytes_bytes_codecs,
+            batch_size=batch_size or config.get("codec_pipeline.batch_size"),
+        )
+
+    @property
+    def supports_partial_decode(self) -> bool:
+        return (len(self.array_array_codecs) + len(self.bytes_bytes_codecs)) == 0 and isinstance(
+            self.array_bytes_codec, ArrayBytesCodecPartialDecodeMixin
+        )
+
+    @property
+    def supports_partial_encode(self) -> bool:
+        return (len(self.array_array_codecs) + len(self.bytes_bytes_codecs)) == 0 and isinstance(
+            self.array_bytes_codec, ArrayBytesCodecPartialEncodeMixin
+        )
+
+    def __iter__(self) -> Iterator[Codec]:
+        for aa_codec in self.array_array_codecs:
+            yield aa_codec
+
+        yield self.array_bytes_codec
+
+        for bb_codec in self.bytes_bytes_codecs:
+            yield bb_codec
+
+    def validate(self, array_metadata: ArrayMetadata) -> None:
+        for codec in self:
+            codec.validate(array_metadata)
+
+    def compute_encoded_size(self, byte_length: int, array_spec: ArraySpec) -> int:
+        for codec in self:
+            byte_length = codec.compute_encoded_size(byte_length, array_spec)
+            array_spec = codec.resolve_metadata(array_spec)
+        return byte_length
+
+    def _codecs_with_resolved_metadata_batched(
+        self, chunk_specs: Iterable[ArraySpec]
+    ) -> tuple[
+        list[tuple[ArrayArrayCodec, list[ArraySpec]]],
+        tuple[ArrayBytesCodec, list[ArraySpec]],
+        list[tuple[BytesBytesCodec, list[ArraySpec]]],
+    ]:
+        aa_codecs_with_spec: list[tuple[ArrayArrayCodec, list[ArraySpec]]] = []
+        chunk_specs = list(chunk_specs)
+        for aa_codec in self.array_array_codecs:
+            aa_codecs_with_spec.append((aa_codec, chunk_specs))
+            chunk_specs = [aa_codec.resolve_metadata(chunk_spec) for chunk_spec in chunk_specs]
+
+        ab_codec_with_spec = (self.array_bytes_codec, chunk_specs)
+        chunk_specs = [
+            self.array_bytes_codec.resolve_metadata(chunk_spec) for chunk_spec in chunk_specs
+        ]
+
+        bb_codecs_with_spec: list[tuple[BytesBytesCodec, list[ArraySpec]]] = []
+        for bb_codec in self.bytes_bytes_codecs:
+            bb_codecs_with_spec.append((bb_codec, chunk_specs))
+            chunk_specs = [bb_codec.resolve_metadata(chunk_spec) for chunk_spec in chunk_specs]
+
+        return (aa_codecs_with_spec, ab_codec_with_spec, bb_codecs_with_spec)
+
+    async def decode_batch(
+        self,
+        chunk_bytes_and_specs: Iterable[tuple[BytesLike | None, ArraySpec]],
+    ) -> Iterable[np.ndarray | None]:
+        chunk_bytes_batch: Iterable[BytesLike | None]
+        chunk_bytes_batch, chunk_specs = _unzip2(chunk_bytes_and_specs)
+
+        (
+            aa_codecs_with_spec,
+            ab_codec_with_spec,
+            bb_codecs_with_spec,
+        ) = self._codecs_with_resolved_metadata_batched(chunk_specs)
+
+        for bb_codec, chunk_spec_batch in bb_codecs_with_spec[::-1]:
+            chunk_bytes_batch = await bb_codec.decode(zip(chunk_bytes_batch, chunk_spec_batch))
+
+        ab_codec, chunk_spec_batch = ab_codec_with_spec
+        chunk_array_batch = await ab_codec.decode(zip(chunk_bytes_batch, chunk_spec_batch))
+
+        for aa_codec, chunk_spec_batch in aa_codecs_with_spec[::-1]:
+            chunk_array_batch = await aa_codec.decode(zip(chunk_array_batch, chunk_spec_batch))
+
+        return chunk_array_batch
+
+    async def decode_partial_batch(
+        self,
+        batch_info: Iterable[tuple[ByteGetter, SliceSelection, ArraySpec]],
+    ) -> Iterable[np.ndarray | None]:
+        assert self.supports_partial_decode
+        assert isinstance(self.array_bytes_codec, ArrayBytesCodecPartialDecodeMixin)
+        return await self.array_bytes_codec.decode_partial(batch_info)
+
+    async def encode_batch(
+        self,
+        chunk_arrays_and_specs: Iterable[tuple[np.ndarray | None, ArraySpec]],
+    ) -> Iterable[BytesLike | None]:
+        chunk_array_batch: Iterable[np.ndarray | None]
+        chunk_specs: Iterable[ArraySpec]
+        chunk_array_batch, chunk_specs = _unzip2(chunk_arrays_and_specs)
+
+        for aa_codec in self.array_array_codecs:
+            chunk_array_batch = await aa_codec.encode(zip(chunk_array_batch, chunk_specs))
+            chunk_specs = resolve_batched(aa_codec, chunk_specs)
+
+        chunk_bytes_batch = await self.array_bytes_codec.encode(zip(chunk_array_batch, chunk_specs))
+        chunk_specs = resolve_batched(self.array_bytes_codec, chunk_specs)
+
+        for bb_codec in self.bytes_bytes_codecs:
+            chunk_bytes_batch = await bb_codec.encode(zip(chunk_bytes_batch, chunk_specs))
+            chunk_specs = resolve_batched(bb_codec, chunk_specs)
+
+        return chunk_bytes_batch
+
+    async def encode_partial_batch(
+        self,
+        batch_info: Iterable[tuple[ByteSetter, np.ndarray, SliceSelection, ArraySpec]],
+    ) -> None:
+        assert self.supports_partial_encode
+        assert isinstance(self.array_bytes_codec, ArrayBytesCodecPartialEncodeMixin)
+        await self.array_bytes_codec.encode_partial(batch_info)
+
+    async def read_batch(
+        self,
+        batch_info: Iterable[tuple[ByteGetter, ArraySpec, SliceSelection, SliceSelection]],
+        out: np.ndarray,
+    ) -> None:
+        if self.supports_partial_decode:
+            chunk_array_batch = await self.decode_partial_batch(
+                [
+                    (byte_getter, chunk_selection, chunk_spec)
+                    for byte_getter, chunk_spec, chunk_selection, _ in batch_info
+                ]
+            )
+            for chunk_array, (_, chunk_spec, _, out_selection) in zip(
+                chunk_array_batch, batch_info
+            ):
+                if chunk_array is not None:
+                    out[out_selection] = chunk_array
+                else:
+                    out[out_selection] = chunk_spec.fill_value
+        else:
+            chunk_bytes_batch = await concurrent_map(
+                [(byte_getter,) for byte_getter, _, _, _ in batch_info],
+                lambda byte_getter: byte_getter.get(),
+                config.get("async.concurrency"),
+            )
+            chunk_array_batch = await self.decode_batch(
+                [
+                    (chunk_bytes, chunk_spec)
+                    for chunk_bytes, (_, chunk_spec, _, _) in zip(chunk_bytes_batch, batch_info)
+                ],
+            )
+            for chunk_array, (_, chunk_spec, chunk_selection, out_selection) in zip(
+                chunk_array_batch, batch_info
+            ):
+                if chunk_array is not None:
+                    tmp = chunk_array[chunk_selection]
+                    out[out_selection] = tmp
+                else:
+                    out[out_selection] = chunk_spec.fill_value
+
+    async def write_batch(
+        self,
+        batch_info: Iterable[tuple[ByteSetter, ArraySpec, SliceSelection, SliceSelection]],
+        value: np.ndarray,
+    ) -> None:
+        if self.supports_partial_encode:
+            await self.encode_partial_batch(
+                [
+                    (byte_setter, value[out_selection], chunk_selection, chunk_spec)
+                    for byte_setter, chunk_spec, chunk_selection, out_selection in batch_info
+                ],
+            )
+
+        else:
+            # Read existing bytes if not total slice
+            async def _read_key(byte_setter: ByteSetter | None) -> BytesLike | None:
+                if byte_setter is None:
+                    return None
+                return await byte_setter.get()
+
+            chunk_bytes_batch: Iterable[BytesLike | None]
+            chunk_bytes_batch = await concurrent_map(
+                [
+                    (None if is_total_slice(chunk_selection, chunk_spec.shape) else byte_setter,)
+                    for byte_setter, chunk_spec, chunk_selection, _ in batch_info
+                ],
+                _read_key,
+                config.get("async.concurrency"),
+            )
+            chunk_array_batch = await self.decode_batch(
+                [
+                    (chunk_bytes, chunk_spec)
+                    for chunk_bytes, (_, chunk_spec, _, _) in zip(chunk_bytes_batch, batch_info)
+                ],
+            )
+
+            def _merge_chunk_array(
+                existing_chunk_array: np.ndarray | None,
+                new_chunk_array_slice: np.ndarray,
+                chunk_spec: ArraySpec,
+                chunk_selection: SliceSelection,
+            ) -> np.ndarray:
+                if is_total_slice(chunk_selection, chunk_spec.shape):
+                    return new_chunk_array_slice
+                if existing_chunk_array is None:
+                    chunk_array = np.empty(
+                        chunk_spec.shape,
+                        dtype=chunk_spec.dtype,
+                    )
+                    chunk_array.fill(chunk_spec.fill_value)
+                else:
+                    chunk_array = existing_chunk_array.copy()  # make a writable copy
+                chunk_array[chunk_selection] = new_chunk_array_slice
+                return chunk_array
+
+            chunk_array_batch = [
+                _merge_chunk_array(chunk_array, value[out_selection], chunk_spec, chunk_selection)
+                for chunk_array, (_, chunk_spec, chunk_selection, out_selection) in zip(
+                    chunk_array_batch, batch_info
+                )
+            ]
+
+            chunk_array_batch = [
+                None if np.all(chunk_array == chunk_spec.fill_value) else chunk_array
+                for chunk_array, (_, chunk_spec, _, _) in zip(chunk_array_batch, batch_info)
+            ]
+
+            chunk_bytes_batch = await self.encode_batch(
+                [
+                    (chunk_array, chunk_spec)
+                    for chunk_array, (_, chunk_spec, _, _) in zip(chunk_array_batch, batch_info)
+                ],
+            )
+
+            async def _write_key(byte_setter: ByteSetter, chunk_bytes: BytesLike | None) -> None:
+                if chunk_bytes is None:
+                    await byte_setter.delete()
+                else:
+                    await byte_setter.set(chunk_bytes)
+
+            await concurrent_map(
+                [
+                    (byte_setter, chunk_bytes)
+                    for chunk_bytes, (byte_setter, _, _, _) in zip(chunk_bytes_batch, batch_info)
+                ],
+                _write_key,
+                config.get("async.concurrency"),
+            )
+
+    async def decode(
+        self,
+        chunk_bytes_and_specs: Iterable[tuple[BytesLike | None, ArraySpec]],
+    ) -> Iterable[np.ndarray | None]:
+        output: list[np.ndarray | None] = []
+        for batch_info in batched(chunk_bytes_and_specs, self.batch_size):
+            output.extend(await self.decode_batch(batch_info))
+        return output
+
+    async def decode_partial(
+        self,
+        batch_info: Iterable[tuple[ByteGetter, SliceSelection, ArraySpec]],
+    ) -> Iterable[np.ndarray | None]:
+        output: list[np.ndarray | None] = []
+        for single_batch_info in batched(batch_info, self.batch_size):
+            output.extend(await self.decode_partial_batch(single_batch_info))
+        return output
+
+    async def encode(
+        self,
+        chunk_arrays_and_specs: Iterable[tuple[np.ndarray | None, ArraySpec]],
+    ) -> Iterable[BytesLike | None]:
+        output: list[BytesLike | None] = []
+        for single_batch_info in batched(chunk_arrays_and_specs, self.batch_size):
+            output.extend(await self.encode_batch(single_batch_info))
+        return output
+
+    async def encode_partial(
+        self,
+        batch_info: Iterable[tuple[ByteSetter, np.ndarray, SliceSelection, ArraySpec]],
+    ) -> None:
+        for single_batch_info in batched(batch_info, self.batch_size):
+            await self.encode_partial_batch(single_batch_info)
+
+    async def read(
+        self,
+        batch_info: Iterable[tuple[ByteGetter, ArraySpec, SliceSelection, SliceSelection]],
+        out: np.ndarray,
+    ) -> None:
+        await concurrent_map(
+            [
+                (single_batch_info, out)
+                for single_batch_info in batched(batch_info, self.batch_size)
+            ],
+            self.read_batch,
+            config.get("async.concurrency"),
+        )
+
+    async def write(
+        self,
+        batch_info: Iterable[tuple[ByteSetter, ArraySpec, SliceSelection, SliceSelection]],
+        value: np.ndarray,
+    ) -> None:
+        await concurrent_map(
+            [
+                (single_batch_info, value)
+                for single_batch_info in batched(batch_info, self.batch_size)
+            ],
+            self.write_batch,
+            config.get("async.concurrency"),
+        )
