--- src/zarr/codecs/pipeline/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from zarr.codecs.pipeline.core import CodecPipeline  # noqa: F401
-from zarr.codecs.pipeline.hybrid import HybridCodecPipeline  # noqa: F401
-from zarr.codecs.pipeline.batched import BatchedCodecPipeline  # noqa: F401
